# pkg/swagger/docs/swagger.yaml
definitions:
  model.LoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.Response:
    properties:
      data: {}
      error: {}
      log_id:
        type: string
      message: {}
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: get the status of server.
      parameters:
      - description: model.LoginPayload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Response'
      summary: User Administrator Login
      tags:
      - Admin - Login
  /api/v1/order/create:
    post:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Create Order
      tags:
      - Order Management System - Order
  /api/v1/order/delete/{param}:
    delete:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: order id
        in: path
        name: param
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Delete the User
      tags:
      - Order Management System - Order
  /api/v1/order/detail/{param}:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: User Id
        in: path
        name: param
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Detail of User
      tags:
      - Order Management System - Order
  /api/v1/order/list:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Get List Order
      tags:
      - Order Management System - Order
  /api/v1/order/update/{param}:
    put:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: User Id
        in: path
        name: param
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Update User
      tags:
      - Order Management System - Order
  /api/v1/product/create:
    post:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Create Product
      tags:
      - Product Management - Product
  /api/v1/product/delete/{param}:
    delete:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: Product Id
        in: path
        name: param
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Delete Product
      tags:
      - Product Management - Product
  /api/v1/product/detail/{param}:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: Product Id
        in: path
        name: param
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Detail of Produt
      tags:
      - Product Management - Product
  /api/v1/product/list:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Get List Product
      tags:
      - Product Management - Product
  /api/v1/product/update/{param}:
    put:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: Product Id
        in: path
        name: param
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Update Product
      tags:
      - Product Management - Product
  /api/v1/user/delete/{param}:
    delete:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: User Id
        in: path
        name: param
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Delete the User
      tags:
      - Identity Access Management - User
  /api/v1/user/detail/{param}:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: User Id
        in: path
        name: param
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Detail of User
      tags:
      - Identity Access Management - User
  /api/v1/user/list:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Get List User
      tags:
      - Identity Access Management - User
  /api/v1/user/register:
    post:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Register User
      tags:
      - Identity Access Management - User
  /api/v1/user/update/{param}:
    put:
      consumes:
      - '*/*'
      description: get the status of server.
      parameters:
      - description: User Id
        in: path
        name: param
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Update User
      tags:
      - Identity Access Management - User
  /migration:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: interface
      summary: Migration
      tags:
      - System Server
swagger: "2.0"